+++ src/components/sdk/MeetingControls/index.tsx
@@ -5,6 +5,7 @@ import AudioInputControl from './AudioInputControl';
 import AudioInputVFControl from './AudioInputVFControl';
 import AudioOutputControl from './AudioOutputControl';
 import ContentShareControl from './ContentShareControl';
+import VideoInputBackgroundBlurControl from './VideoInputBackgroundBlurControl';
 import VideoInputControl from './VideoInputControl';
 
 export {
@@ -12,5 +13,6 @@ export {
   AudioInputVFControl,
   AudioOutputControl,
   VideoInputControl,
+  VideoInputBackgroundBlurControl,
   ContentShareControl,
 };